Minor tweaks
(1) It turns out Windows didn't like command line arguments enclosed in single
quotes, but was okay with arguments in double quotes.  This came up when some
students went to copy-and-paste the git command in the README to commit their
changes.

(2) Near ":smile:", there was a smart quote that accidentally got pasted into
the README rather than a straight quote.